feat: make finished the flow to validator author

+++ libraryapi/src/main/java/io/github/simaomenezes/libraryapi/service/AuthorService.java
@@ -1,7 +1,10 @@
 package io.github.simaomenezes.libraryapi.service;
 
+import io.github.simaomenezes.libraryapi.exceptions.OperationNotAllowException;
 import io.github.simaomenezes.libraryapi.model.Author;
 import io.github.simaomenezes.libraryapi.repository.AuthorRepository;
+import io.github.simaomenezes.libraryapi.repository.BookRepository;
+import io.github.simaomenezes.libraryapi.validator.AuthorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AuthorService {
 
     private final AuthorRepository repository;
+    private final AuthorValidator validator;
+    private final BookRepository bookRepository;
 
-    public AuthorService(AuthorRepository repository) {
+    public AuthorService(AuthorRepository repository, AuthorValidator validator, BookRepository bookRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.bookRepository = bookRepository;
     }
 
     public Author add(Author author){
+        validator.validator(author);
         return repository.save(author);
     }
 
@@ -25,6 +33,7 @@ public class AuthorService {
         if(author.getId() == null){
             throw new IllegalArgumentException("The to save, the author need on saved the bd data");
         }
+        validator.validator(author);
         repository.saveAndFlush(author);
     }
 
@@ -33,7 +42,14 @@ public class AuthorService {
     }
 
     public void delete(Author author){
+        if(hashBook(author)){
+            throw new OperationNotAllowException("Can't delete author with a book created!");
+        }
         repository.delete(author);
+    }
+
+    private boolean hashBook(Author author) {
+        return bookRepository.existsByAuthor(author);
     }
 
     public List<Author> search(String name, String nationality){
